+++ Cmf.Custom.Tests/Cmf.Custom.Tests.Biz.Common/ERP/ERPMessageSerializer.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
+using System.Xml;
+using System.Xml.Serialization;
+
+namespace Cmf.Custom.Tests.Biz.Common.ERP
+{
+    public class ERPMessageSerializer<T> where T : class
+    {
+        public static string Serialize(T obj)
+        {
+            XmlSerializer xsSubmit = new XmlSerializer(typeof(T));
+            using (var sww = new StringWriter())
+            {
+                using (XmlTextWriter writer = new XmlTextWriter(sww) { Formatting = Formatting.Indented })
+                {
+                    xsSubmit.Serialize(writer, obj);
+                    return sww.ToString();
+                }
+            }
+        }
+    }
+}