+++ Cmf.Custom.Tests/Cmf.Custom.Tests.Biz/NiceLabelPrinting/CustomGetNiceLabelPrintingInformation.cs
@@ -1,18 +1,17 @@
 ﻿using Cmf.Custom.Tests.Biz.Common;
 using Cmf.Custom.TestUtilities;
-using Cmf.Foundation.BusinessOrchestration.DynamicExecutionEngineManagement.InputObjects;
-using Cmf.Foundation.BusinessOrchestration.DynamicExecutionEngineManagement.OutputObjects;
 using Cmf.Navigo.BusinessObjects;
 using Cmf.Navigo.BusinessOrchestration.ContainerManagement.InputObjects;
 using Cmf.Navigo.BusinessOrchestration.ContainerManagement.OutputObjects;
 using Cmf.Navigo.BusinessOrchestration.FacilityManagement.FlowManagement.InputObjects;
 using Cmf.Navigo.BusinessOrchestration.MaterialManagement.InputObjects;
-using Cmf.Navigo.BusinessOrchestration.MaterialManagement.OutputObjects;
 using Cmf.TestScenarios.Others;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
-using System.Text;
+using System.IO;
+using System.Linq;
+using System.Threading;
 
 namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
 {
@@ -43,23 +42,25 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
         /// <summary>
         /// Description:
         ///     - Fill custom Smart Table CustomMaterialNiceLabelPrintContext with information and isEnabled = true
-        ///     - Track in and track out material
+        ///     - Track in and track out lot
         ///     
         /// Acceptance Criteria:
-        ///     - Resolve the Smart Table information 
+        ///     - Resolve the Smart Table information
+        ///     - Valida the file contents
         ///     
         /// </summary>
         /// <TestCaseID>CustomGetNiceLabelPrintingInformation.CustomGetNiceLabelPrintingInformation_TrackOutMaterial_HappyPath</TestCaseID>
         /// <Author>David Guilherme</Author>
-        // [TestMethod]
+        //[TestMethod]
         public void CustomGetNiceLabelPrintingInformation_TrackOutMaterial_HappyPath()
         {
+            int subMaterialQuantity = 2;
 
             #region Setup
 
             #region Create Material
-            ///<Step> Create a material </Step>
-            Material material = new Material()
+            ///<Step> Create a lot </Step>
+            Material lot = new Material()
             {
                 Name = Guid.NewGuid().ToString("N"),
                 Facility = GenericGetsScenario.GetObjectByName<Facility>(AMSOsramConstants.DefaultFacilityName),
@@ -69,12 +70,37 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
                 {
                     SequenceFlowPath = AMSOsramConstants.DefaultFlowName + ":1/" + AMSOsramConstants.DefaultMVPStepName + @":30"
                 }.GetCorrelationIdFlowPathSync().CorrelationIdFlowPath,
-                Form = AMSOsramConstants.DefaultMaterialLogisticalWaferForm,
+                Form = AMSOsramConstants.DefaultMaterialFormName,
                 PrimaryQuantity = 10,
                 PrimaryUnits = AMSOsramConstants.DefaultMaterialUnit
             };
-            material.Create();
-            customTeardownManager.Push(material);
+            lot.Create();
+            customTeardownManager.Push(lot);
+
+            //Product 
+            Product product = GenericGetsScenario.GetObjectByName<Product>(AMSOsramConstants.DefaultProductName);
+            // Expand MainMaterial to create 3 sub-materials
+            MaterialCollection subMaterials = new MaterialCollection();
+            for (var i = 0; i < subMaterialQuantity; i++)
+            {
+                // Create Sub Material to expand for main lot
+                var wafer = new Material()
+                {
+                    PrimaryQuantity = 1,
+                    Product = product
+                };
+                subMaterials.Add(wafer);
+            }
+
+            ExpandMaterialInput input = new ExpandMaterialInput()
+            {
+                Material = lot,
+                SubMaterials = subMaterials,
+                Form = AMSOsramConstants.DefaultMaterialLogisticalWaferForm
+            };
+
+            subMaterials = input.ExpandMaterialSync().ExpandedSubMaterials;
+
             #endregion
 
             #region SmartTable
@@ -90,7 +116,7 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
             {
                     { "Step", AMSOsramConstants.DefaultMVPStepName },
                     { "Operation", "TrackOut" },
-                    { "Material", material.Name },
+                    { "Material", lot.Name },
                     { "Printer", expectedPrinter },
                     { "Label", expectedLabel },
                     { "Quantity", expectedQuantity.ToString() },
@@ -103,83 +129,108 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
 
             Container container = GenericGetsScenario.GetObjectByName<Container>(AMSOsramConstants.DefaultContainerName);
             container.Load();
+            int position = 1;
 
-            MaterialContainer materialContainerRelation = new MaterialContainer()
-            {
-                SourceEntity = material,
-                TargetEntity = container,
-                Position = 1
-            };
             MaterialContainerCollection materialContainerCollection = new MaterialContainerCollection();
-            materialContainerCollection.Add(materialContainerRelation);
+            foreach (Material subMaterial in subMaterials)
+            {
+                MaterialContainer materialContainerRelation = new MaterialContainer()
+                {
+                    SourceEntity = subMaterial,
+                    TargetEntity = container,
+                    Position = position
+                };
+                position ++;
+                materialContainerCollection.Add(materialContainerRelation);
+            }
+
             AssociateMaterialsWithContainerInput associateInput = new AssociateMaterialsWithContainerInput();
             associateInput.MaterialContainerRelations = materialContainerCollection;
             associateInput.Container = container;
 
-            AssociateMaterialsWithContainerOutput associateOutput = new AssociateMaterialsWithContainerOutput();
-            associateOutput = associateInput.AssociateMaterialsWithContainerSync();
+            AssociateMaterialsWithContainerOutput associateOutput = associateInput.AssociateMaterialsWithContainerSync();
+
             #endregion
+
+            Resource printer = GenericGetsScenario.GetObjectByName<Resource>("NiceLabelPrinter");
+            printer.Load();
+            printer.LoadAttributes(new string[] { "AutomationEquipmentAddress" });
+
+            Resource processResource = GenericGetsScenario.GetObjectByName<Resource>("LOWS0101");
+            processResource.Load();
+
+            Area area = processResource.Area;
+            area.Load();
 
             #endregion
 
             ///<Step> Track In Material </Step>
-            material.ComplexDispatchAndTrackIn();
-            material.Load(1);
+            lot.Load();
+            lot.ComplexDispatchAndTrackIn(processResource);
+            
+            ///<Step> Execute Trackout </Step>
+            lot.ComplexTrackOutMaterial();
 
-            ///<Step> Execute DEE with Trackout information</Step>
-            #region Execute DEE
-            Foundation.Common.DynamicExecutionEngine.Action customNiceLabelPrint = DEEUtilities.GetActionByName(AMSOsramConstants.CustomNiceLabelPrintDEE);
 
-            Dictionary<Material, ComplexTrackOutParameters> trackOutMaterials = new Dictionary<Material, ComplexTrackOutParameters>()
+            ///<Step> Get Generated File for Nice Label Printing </Step> 
+            Thread.Sleep(1000);
+            string directoryPath = printer.Attributes["AutomationEquipmentAddress"].ToString();
+            FileInfo file = new DirectoryInfo(directoryPath).GetFiles().FirstOrDefault(f => f.Name.Contains(lot.Name));
+            Assert.IsTrue(file != null, $"Generated file for label printing should container the lot name.");
+            
+            System.Text.Encoding encoding = System.Text.Encoding.GetEncoding("ISO-8859-1");
+            string text;
+            using (System.IO.StreamReader sr = new System.IO.StreamReader(file.FullName, encoding))
             {
-                {material,new ComplexTrackOutParameters() }
-            };
-
-            ComplexTrackOutMaterialsInput input = new ComplexTrackOutMaterialsInput()
-            {
-                Material = trackOutMaterials
-            };
-
-            ExecuteActionOutput output = new ExecuteActionInput
-            {
-                Action = customNiceLabelPrint,
-                Input = new Dictionary<string, object>()
-                {
-                    {"ComplexTrackOutMaterialsInput", input}
-                }
-            }.ExecuteActionSync(); 
-            #endregion
-
-            Dictionary<string, object> outputInformation = (Dictionary<string, object>)output.Output["NiceLabelInformation"];
-
-            ///<ExpectedValue> The DEE should return the correct information filled on the smart table </ExpectedValue>
-            foreach (string materialName in outputInformation.Keys)
-            {
-
-                // TODO ::: FIX THIS TEST
-
-
-                Dictionary<string, object> materialInformationToPrint = (Dictionary<string, object>)outputInformation[materialName];
-                Assert.IsTrue(materialInformationToPrint["LotName"].ToString().Equals(materialName), $"Column LotName should have the value: {materialName}, instead is: {materialInformationToPrint["LotName"]}.");
-                Assert.IsTrue(materialInformationToPrint["PRINTER_NAME"].ToString().Equals(expectedPrinter), $"Column Printer should have the value: {expectedPrinter}, instead is: {materialInformationToPrint["PRINTER_NAME"]}.");
-                Assert.IsTrue(materialInformationToPrint["LABEL_NAME"].ToString().Equals(expectedLabel), $"Column Printer should have the value: {expectedLabel}, instead is: {materialInformationToPrint["LABEL_NAME"]}.");
-                Assert.IsTrue(materialInformationToPrint["LABEL_QUANTITY"].ToString().Equals(expectedQuantity), $"Column Printer should have the value: {expectedQuantity}, instead is: {materialInformationToPrint["LABEL_QUANTITY"]}.");
-                Assert.IsTrue(materialInformationToPrint["ProductName"].ToString().Equals(AMSOsramConstants.DefaultProductName), $"Column Product should have the value: {AMSOsramConstants.DefaultProductName}, instead is: {materialInformationToPrint["ProductName"]}.");
-                
-
-                Assert.IsTrue(string.IsNullOrEmpty(materialInformationToPrint["ProductGroupName"]?.ToString()), $"Column Product Group should be empty. '" + materialInformationToPrint["ProductGroupName"]?.ToString() + "'");
-                Assert.IsTrue(materialInformationToPrint["ProductDesc"].ToString().Equals(material.Product?.Description), $"Column Product Description should have the value: {material.Product?.Description}, instead is: {materialInformationToPrint["ProductDesc"]}.");
-                Assert.IsTrue(materialInformationToPrint["ProductType"].ToString().Equals(material.Product?.ProductType.ToString()), $"Column Product Product Type should have the value: {material.Product?.ProductType.ToString()}, instead is: {materialInformationToPrint["ProductType"]}.");
-                Assert.IsTrue(materialInformationToPrint["Product_Type"].ToString().Equals(material.Product?.Type), $"Column Product Type should have the value: {material.Product?.Type}, instead is: {materialInformationToPrint["Product_Type"]}.");
-                Assert.IsTrue(materialInformationToPrint["FlowName"].ToString().Equals(material.Flow.Name), $"Column Flow Name should have the value: {material.Flow.Name}, instead is: {materialInformationToPrint["FlowName"]}.");
-                Assert.IsTrue(materialInformationToPrint["ContainerName"].ToString().Equals(AMSOsramConstants.DefaultContainerName), $"Column Container Name should have the value: {AMSOsramConstants.DefaultContainerName}, instead is: {materialInformationToPrint["ContainerName"]}.");
-                Assert.IsTrue(materialInformationToPrint["ResourceName"].ToString().Equals(material.LastProcessedResource.Name), $"Column Resource Name should have the value: {material.LastProcessedResource.Name}, instead is: {materialInformationToPrint["ResourceName"]}.");
-
-                Assert.IsTrue(string.Format("{0:0.##}", Convert.ToDecimal(materialInformationToPrint["LotPrimaryQty"])).Equals("10"), $"Column Lot Primary Quantity should have the value: 10, instead is: {string.Format("{0:0.##}", Convert.ToDecimal(materialInformationToPrint["LotPrimaryQty"]))}.");
-                Assert.IsTrue(string.IsNullOrEmpty(materialInformationToPrint["LotSecundaryQty"]?.ToString()), $"Column Lot Secondary Quantity should have the value: {material.LastProcessedResource.Name}, instead is: {materialInformationToPrint["LotSecundaryQty"]}.");
-                Assert.IsTrue(materialInformationToPrint["Lot_Type"].ToString().Equals(material.Type), $"Column Lot Type should have the value: {material.Type}, instead is: {materialInformationToPrint["Lot_Type"]}.");
+                text = sr.ReadToEnd();
             }
 
+            string[] lines = text.Split(
+                new string[] { "\r\n", "\r", "\n" },
+                StringSplitOptions.None
+            );
+
+            lot.Product.Load();
+            lot.Product.ProductGroup.Load();
+
+            ///<ExpectedValue> The DEE should return the correct information filled on the smart table </ExpectedValue>
+            Dictionary<string, string> expectedFileValues = new Dictionary<string, string>()
+            {
+                { "LABEL_NAME", expectedLabel },
+                { "PRINTER_NAME", expectedPrinter },
+                { "LABEL_QUANTITY", expectedQuantity },
+                { "LotName", lot.Name },
+                { "LotAlias", string.Empty },
+                { "ProductName", AMSOsramConstants.DefaultProductName },
+                { "ProductDesc", lot.Product.Description },
+                { "ProductType", lot.Product.ProductType.ToString() },
+                { "Product_Type", lot.Product.Type },
+                { "ProductGroupName", lot.Product.ProductGroup.Name },
+                { "ProductGroup_Type", lot.Product.ProductGroup.Type },
+                { "FlowName", AMSOsramConstants.DefaultFlowName },
+                { "BatchName", string.Empty },
+                { "ContainerName", AMSOsramConstants.DefaultContainerName },
+                { "ExperimentName", string.Empty },
+                { "ProductionOrder", string.Empty },
+                { "LotOwner", string.Empty },
+                { "ResourceName", processResource.Name },
+                { "LotWaferCount", string.Empty },
+                { "LotPrimaryQty", "8" },
+                { "LotSecondaryQty", string.Empty },
+                { "Lot_Type", lot.Type },
+                { "Area", area.Name },
+                { "Facility", AMSOsramConstants.DefaultFacilityName }
+            };
+
+            foreach (string line in lines)
+            {
+                if (!line.Equals("END"))
+                {
+                    string column = line.Split("=")[0];
+                    string value = column.Equals("LotPrimaryQty") ? string.Format("{0:0.##}", Convert.ToDecimal(line.Split("=")[1])) : line.Split("=")[1];
+                    Assert.IsTrue(value.Equals(expectedFileValues[column]), $"Column {column} should have the value: {expectedFileValues[column]}, but instead has the value: {value}");
+                }
+            }
         }
 
         /// <summary>
@@ -200,8 +251,8 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
             #region Setup
 
             #region Create Material
-            ///<Step> Create a material </Step>
-            Material material = new Material()
+            ///<Step> Create a lot </Step>
+            Material lot = new Material()
             {
                 Name = Guid.NewGuid().ToString("N"),
                 Facility = GenericGetsScenario.GetObjectByName<Facility>(AMSOsramConstants.DefaultFacilityName),
@@ -215,51 +266,33 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
                 PrimaryQuantity = 10,
                 PrimaryUnits = AMSOsramConstants.DefaultMaterialUnit
             };
-            material.Create();
-            customTeardownManager.Push(material);
+            lot.Create();
+            customTeardownManager.Push(lot);
             #endregion
 
             ///<Step> Ensure Smart Table CustomMaterialNiceLabelPrintContext is empty</Step>
             smartTableManager.ClearSmartTable(AMSOsramConstants.CustomMaterialNiceLabelPrintContextSmartTable);
-            
+
+
+            Resource printer = GenericGetsScenario.GetObjectByName<Resource>("NiceLabelPrinter");
+            printer.Load();
+            printer.LoadAttributes(new string[] { "AutomationEquipmentAddress" });
+
             #endregion
 
             ///<Step> Track In Material </Step>
-            material.ComplexDispatchAndTrackIn();
-            material.Load(1);
+            lot.ComplexDispatchAndTrackIn();
+            lot.Load();
 
-            ///<Step> Execute DEE with Trackout information</Step>
-            #region Execute DEE
-            Foundation.Common.DynamicExecutionEngine.Action customNiceLabelPrint = DEEUtilities.GetActionByName(AMSOsramConstants.CustomNiceLabelPrintDEE);
+            ///<Step> Execute Trackout </Step>
+            lot.ComplexTrackOutMaterial();
 
-            Dictionary<Material, ComplexTrackOutParameters> trackOutMaterials = new Dictionary<Material, ComplexTrackOutParameters>()
-            {
-                {material,new ComplexTrackOutParameters() }
-            };
 
-            ComplexTrackOutMaterialsInput input = new ComplexTrackOutMaterialsInput()
-            {
-                Material = trackOutMaterials
-            };
-
-            ExecuteActionOutput output = new ExecuteActionInput
-            {
-                Action = customNiceLabelPrint,
-                Input = new Dictionary<string, object>()
-                {
-                    {"ComplexTrackOutMaterialsInput", input}
-                }
-            }.ExecuteActionSync();
-            #endregion
-
-            Dictionary<string, object> outputInformation = (Dictionary<string, object>)output.Output["NiceLabelInformation"];
-
-            ///<ExpectedValue> The DEE should return an empty output </ExpectedValue>
-            foreach (string materialName in outputInformation.Keys)
-            {
-                Dictionary<string, object> materialInformationToPrint = (Dictionary<string, object>)outputInformation[materialName];
-                Assert.IsTrue(materialInformationToPrint.Count == 0, $"Output should be empty.");
-            }
+            ///<Step> Get Generated File for Nice Label Printing </Step> 
+            Thread.Sleep(1000);
+            string directoryPath = printer.Attributes["AutomationEquipmentAddress"].ToString();
+            FileInfo file = new DirectoryInfo(directoryPath).GetFiles().FirstOrDefault(f => f.Name.Contains(lot.Name));
+            Assert.IsTrue(file == null, $"Generated file for label printing should container the lot name.");
            
         }
 
@@ -280,8 +313,8 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
             #region Setup
 
             #region Create Material
-            ///<Step> Create a material </Step>
-            Material material = new Material()
+            ///<Step> Create a lot </Step>
+            Material lot = new Material()
             {
                 Name = Guid.NewGuid().ToString("N"),
                 Facility = GenericGetsScenario.GetObjectByName<Facility>(AMSOsramConstants.DefaultFacilityName),
@@ -295,8 +328,8 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
                 PrimaryQuantity = 10,
                 PrimaryUnits = AMSOsramConstants.DefaultMaterialUnit
             };
-            material.Create();
-            customTeardownManager.Push(material);
+            lot.Create();
+            customTeardownManager.Push(lot);
             #endregion
 
             #region SmartTable
@@ -312,7 +345,7 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
             {
                     { "Step", AMSOsramConstants.DefaultMVPStepName },
                     { "Operation", "TrackOut" },
-                    { "Material", material.Name },
+                    { "Material", lot.Name },
                     { "Printer", expectedPrinter },
                     { "Label", expectedLabel },
                     { "Quantity", expectedQuantity.ToString() },
@@ -321,44 +354,25 @@ namespace Cmf.Custom.Tests.Biz.NiceLabelPrinting
 
             #endregion
 
+            Resource printer = GenericGetsScenario.GetObjectByName<Resource>("NiceLabelPrinter");
+            printer.Load();
+            printer.LoadAttributes(new string[] { "AutomationEquipmentAddress" });
+
             #endregion
 
             ///<Step> Track In Material </Step>
-            material.ComplexDispatchAndTrackIn();
-            material.Load(1);
+            lot.ComplexDispatchAndTrackIn();
+            lot.Load();
 
-            ///<Step> Execute DEE with Trackout information</Step>
-            #region Execute DEE
-            Foundation.Common.DynamicExecutionEngine.Action customNiceLabelPrint = DEEUtilities.GetActionByName(AMSOsramConstants.CustomNiceLabelPrintDEE);
+            ///<Step> Execute Trackout </Step>
+            lot.ComplexTrackOutMaterial();
 
-            Dictionary<Material, ComplexTrackOutParameters> trackOutMaterials = new Dictionary<Material, ComplexTrackOutParameters>()
-            {
-                {material,new ComplexTrackOutParameters() }
-            };
 
-            ComplexTrackOutMaterialsInput input = new ComplexTrackOutMaterialsInput()
-            {
-                Material = trackOutMaterials
-            };
-
-            ExecuteActionOutput output = new ExecuteActionInput
-            {
-                Action = customNiceLabelPrint,
-                Input = new Dictionary<string, object>()
-                {
-                    {"ComplexTrackOutMaterialsInput", input}
-                }
-            }.ExecuteActionSync();
-            #endregion
-
-            Dictionary<string, object> outputInformation = (Dictionary<string, object>)output.Output["NiceLabelInformation"];
-
-            ///<ExpectedValue> The DEE should not return the correct information filled on the smart table </ExpectedValue>
-            foreach (string materialName in outputInformation.Keys)
-            {
-                Dictionary<string, object> materialInformationToPrint = (Dictionary<string, object>)outputInformation[materialName];
-                Assert.IsTrue(materialInformationToPrint.Count == 0, $"Output should be empty.");
-            }
+            ///<Step> Get Generated File for Nice Label Printing </Step> 
+            Thread.Sleep(1000);
+            string directoryPath = printer.Attributes["AutomationEquipmentAddress"].ToString();
+            FileInfo file = new DirectoryInfo(directoryPath).GetFiles().FirstOrDefault(f => f.Name.Contains(lot.Name));
+            Assert.IsTrue(file == null, $"Generated file for label printing should container the lot name.");
         }
 
     }