+++ Cmf.Custom.Tests/Cmf.Custom.Tests.IoT/Tests/EvatecClusterline200II/EvatecClusterline200II.cs
@@ -13,12 +13,12 @@ using AMSOsramEIAutomaticTests.Objects.Extensions;
 using AMSOsramEIAutomaticTests.Objects.Utilities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Cmf.Custom.TestUtilities;
-using Cmf.Custom.Tests.IoT.Tests.EvatecClusterline200II;
 using cmConnect.TestFramework.EquipmentSimulator.Objects;
 using Cmf.Foundation.BusinessOrchestration.QueryManagement.InputObjects;
 using Cmf.Foundation.BusinessOrchestration.QueryManagement.OutputObjects;
 using System.Data;
 using Cmf.Foundation.BusinessObjects.QueryObject;
+using Cmf.Custom.Tests.IoT.Tests.Common;
 
 namespace AMSOsramEIAutomaticTests.EvatecClusterline200II
 {
@@ -719,17 +719,16 @@ namespace AMSOsramEIAutomaticTests.EvatecClusterline200II
             }*/
 
             reply.Item.Clear();
-            var mainList = new SecsItem();
-            mainList.SetTypeToList();
+            var attributeList = new SecsItem();
+            attributeList.SetTypeToList();
 
-            var jobItem = new SecsItem { ASCII = "Job1" };
+            var jobItem = new SecsItem { ASCII = controlJobName };
 
-            mainList.Add(jobItem);
+            //mainList.Add(jobItem);
 
             var reportList = new SecsItem();
             reportList.SetTypeToList();
 
-            mainList.Add(reportList);
 
             var ackList = new SecsItem();
             ackList.SetTypeToList();
@@ -738,6 +737,10 @@ namespace AMSOsramEIAutomaticTests.EvatecClusterline200II
 
             var errorList = new SecsItem();
             errorList.SetTypeToList();
+
+            reportList.Add(ack);
+            reportList.Add(errorList);
+
 
             if (createControlJobDenied)
             {
@@ -752,13 +755,9 @@ namespace AMSOsramEIAutomaticTests.EvatecClusterline200II
                 errorList.Add(error);
             }
 
-            ackList.Add(ack);
-
-            ackList.Add(errorList);
-
-            mainList.Add(ackList);
-
-            reply.Item.Add(mainList);
+            reply.Item.Add(jobItem);
+            reply.Item.Add(attributeList);
+            reply.Item.Add(reportList);
             return true;
         }