+++ Cmf.Custom.Help/src/packages/cmf.docs.area.amsosram/assets/releasenotes.md
@@ -15,5 +15,6 @@ Each sprint package includes the following artifacts:
 List of delivered packages:
 * [AMSOsram-20220215-1.0.0](/AMSOsram/releasenotes>AMSOsram-20220215-1_0_0)
 * [AMSOsram-20220228-1.1.0](/AMSOsram/releasenotes>AMSOsram-20220228-1_1_0)
+* [AMSOsram-20220314-1.2.0](/AMSOsram/releasenotes>AMSOsram-20220314-1.2.0)